requirements.txt
pip~=24.3.1
fastapi~=0.115.6
requests~=2.32.3
cachetools~=5.5.0
python-decouple~=3.8
SQLAlchemy~=2.0.37
pydantic~=2.10.5
alembic~=1.14.0